Review comments addressed  - File cleanup, removal of un-necessary static class.
