+++ Los Santos RED/lsr/Player/Inventory/Items/CraftableItem.cs
@@ -48,13 +48,6 @@ public class CraftableItem
     }
 
 }
-
-public class Ingredient
-{
-    public string IngredientName { get; set; }
-    public int Quantity { get; set; }
-}
-
 public enum CraftableType
 {
     ModItem,
