+++ Los Santos RED/lsr/Player/Inventory/Items/Ingredient.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+public class Ingredient
+{
+    public string IngredientName { get; set; }
+    public int Quantity { get; set; }
+}
